visualDet3D/requirement.txt
pillow
scikit-image
fire
matplotlib
opencv-python
numba
easydict
cython
tqdm
future
coloredlogs